requirement.txt
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
lxml==4.9.2
numpy==1.25.0
outcome==1.2.0
pyee==9.0.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
wsproto==1.2.0
